Move ShaderWriter to Common/GPU, along with some stuff from ShaderCommon.h.
