+++ GPU/Common/FragmentShaderGenerator.cpp
@@ -21,13 +21,13 @@
 #include "Common/Log.h"
 #include "Common/StringUtils.h"
 #include "Common/GPU/OpenGL/GLFeatures.h"
+#include "Common/GPU/ShaderWriter.h"
 #include "Core/Reporting.h"
 #include "Core/Config.h"
 #include "GPU/Common/GPUStateUtils.h"
 #include "GPU/Common/ShaderId.h"
 #include "GPU/Common/ShaderUniforms.h"
 #include "GPU/Common/FragmentShaderGenerator.h"
-#include "GPU/Common/ShaderWriter.h"
 #include "GPU/ge_constants.h"
 #include "GPU/GPUState.h"
