+++ GPU/Common/ShaderCommon.h
@@ -24,18 +24,6 @@ namespace Draw {
 	class DrawContext;
 }
 
-enum ShaderLanguage {
-	GLSL_140,  // really covers a lot more. This set of languages is not good.
-	GLSL_300,
-	GLSL_VULKAN,
-	HLSL_D3D9,
-	HLSL_D3D11,
-};
-
-inline bool ShaderLanguageIsOpenGL(ShaderLanguage lang) {
-	return lang == GLSL_140 || lang == GLSL_300;
-}
-
 enum DebugShaderType {
 	SHADER_TYPE_VERTEX = 0,
 	SHADER_TYPE_FRAGMENT = 1,
@@ -50,11 +38,6 @@ enum DebugShaderStringType {
 	SHADER_STRING_SHORT_DESC = 0,
 	SHADER_STRING_SOURCE_CODE = 1,
 	SHADER_STRING_STATS = 2,
-};
-
-enum class ShaderStage {
-	Vertex,
-	Fragment
 };
 
 // Shared between the backends. Not all are necessarily used by each backend, but this lets us share
@@ -145,28 +128,6 @@ enum DoLightComputation {
 	LIGHT_OFF,
 	LIGHT_SHADE,
 	LIGHT_FULL,
-};
-
-struct ShaderLanguageDesc {
-	explicit ShaderLanguageDesc(ShaderLanguage lang);
-
-	int glslVersionNumber = 0;
-	ShaderLanguage shaderLanguage;
-	bool gles = false;
-	const char *varying_fs = nullptr;
-	const char *varying_vs = nullptr;
-	const char *attribute = nullptr;
-	const char *fragColor0 = nullptr;
-	const char *fragColor1 = nullptr;
-	const char *texture = nullptr;
-	const char *texelFetch = nullptr;
-	const char *lastFragData = nullptr;
-	const char *framebufferFetchExtension = nullptr;
-	const char *vsOutPrefix = "";
-	bool glslES30 = false;
-	bool bitwiseOps = false;
-	bool forceMatrix4x4 = false;
-	bool coefsFromBuffers = false;
 };
 
 // PSP vertex format.
