+++ UWP/GPU_UWP/GPU_UWP.vcxproj.filters
@@ -57,7 +57,6 @@
     <ClCompile Include="..\..\GPU\Software\RasterizerRectangle.cpp" />
     <ClCompile Include="..\..\GPU\Common\FragmentShaderGenerator.cpp" />
     <ClCompile Include="..\..\GPU\Common\VertexShaderGenerator.cpp" />
-    <ClCompile Include="..\..\GPU\Common\ShaderWriter.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\GPU\Common\DepalettizeShaderCommon.h" />
@@ -115,6 +114,5 @@
     <ClInclude Include="..\..\GPU\Software\RasterizerRectangle.h" />
     <ClInclude Include="..\..\GPU\Common\FragmentShaderGenerator.h" />
     <ClInclude Include="..\..\GPU\Common\VertexShaderGenerator.h" />
-    <ClInclude Include="..\..\GPU\Common\ShaderWriter.h" />
   </ItemGroup>
 </Project>
